+++ crates/nixci/src/step/core.rs
@@ -1,10 +1,14 @@
 //! All CI steps available in nixci
 use clap::Parser;
-use nix_rs::{command::NixCmd, flake::url::FlakeUrl};
+use nix_rs::{
+    command::NixCmd,
+    flake::{system::System, url::FlakeUrl},
+};
 use serde::Deserialize;
 
 use super::{
     build::{BuildStep, BuildStepArgs},
+    custom::CustomSteps,
     flake_check::FlakeCheckStep,
     lockfile::LockfileStep,
 };
@@ -27,7 +31,10 @@ pub struct Steps {
     /// [FlakeCheckStep]
     #[serde(default, rename = "flake-check")]
     pub flake_check_step: FlakeCheckStep,
-    // TODO: custom steps
+
+    /// Custom steps
+    #[serde(rename = "custom")]
+    pub custom_steps: CustomSteps,
 }
 
 /// CLI arguments associated with [Steps]
@@ -45,19 +52,28 @@ impl Steps {
         cmd: &NixCmd,
         verbose: bool,
         run_cmd: &RunCommand,
+        systems: &[System],
         url: &FlakeUrl,
         subflake: &SubflakeConfig,
     ) -> anyhow::Result<()> {
-        self.lockfile_step.run(cmd, url, subflake).await?;
+        if self.lockfile_step.enable {
+            self.lockfile_step.run(cmd, url, subflake).await?;
+        }
 
-        let res = self
-            .build_step
-            .run(cmd, verbose, run_cmd, url, subflake)
-            .await?;
-        // TODO: Support --json for structured output grouped by steps
-        res.print();
+        if self.build_step.enable {
+            let res = self
+                .build_step
+                .run(cmd, verbose, run_cmd, url, subflake)
+                .await?;
+            // TODO: Support --json for structured output grouped by steps
+            res.print();
+        }
 
-        self.flake_check_step.run(cmd, url, subflake).await?;
+        if self.flake_check_step.enable {
+            self.flake_check_step.run(cmd, url, subflake).await?;
+        }
+
+        self.custom_steps.run(cmd, systems, url, subflake).await?;
 
         Ok(())
     }